Organize 15m code a bit
