+++ apps/fifteen_min/src/lib.rs
@@ -7,10 +7,12 @@ use widgetry::Settings;
 #[macro_use]
 extern crate log;
 
+mod amenities_details;
 mod bus;
 mod find_amenities;
 mod find_home;
 mod isochrone;
+mod render;
 mod viewer;
 
 type App = map_gui::SimpleApp<()>;
