+++ apps/fifteen_min/src/viewer.rs
@@ -5,23 +5,22 @@
 //! See https://github.com/a-b-street/abstreet/issues/393 for more context.
 
 use abstutil::prettyprint_usize;
-use geom::{Distance, Duration, FindClosest, Percent};
-use map_gui::tools::{draw_isochrone, CityPicker, Navigator};
+use geom::{Distance, FindClosest, Percent};
+use map_gui::tools::{CityPicker, Navigator};
 use map_gui::ID;
 use map_model::connectivity::WalkingOptions;
-use map_model::{AmenityType, Building, BuildingID, LaneType};
+use map_model::{AmenityType, Building, BuildingID};
 use std::str::FromStr;
-use widgetry::table::{Col, Filter, Table};
-use widgetry::tools::{open_browser, ColorLegend, PopupMsg, URLManager};
+use widgetry::tools::{ColorLegend, PopupMsg, URLManager};
 use widgetry::{
     lctrl, Cached, Choice, Color, Drawable, EventCtx, GeomBatch, GfxCtx, HorizontalAlignment, Key,
-    Line, Outcome, Panel, RewriteColor, State, Text, Toggle, Transition, VerticalAlignment, Widget,
+    Line, Outcome, Panel, State, Text, Toggle, Transition, VerticalAlignment, Widget,
 };
 
 use crate::find_amenities::FindAmenity;
 use crate::find_home::FindHome;
 use crate::isochrone::{Isochrone, MovementOptions, Options};
-use crate::App;
+use crate::{render, App};
 
 /// This is the UI state for exploring the isochrone/walkshed from a single building.
 pub struct Viewer {
@@ -53,9 +52,9 @@ impl Viewer {
         };
         let start = app.map.get_b(start);
         let isochrone = Isochrone::new(ctx, app, vec![start.id], options);
-        let highlight_start = draw_star(ctx, start);
+        let highlight_start = render::draw_star(ctx, start);
         let panel = build_panel(ctx, app, start, &isochrone);
-        let draw_unwalkable_roads = draw_unwalkable_roads(ctx, app, &isochrone.options);
+        let draw_unwalkable_roads = render::draw_unwalkable_roads(ctx, app, &isochrone.options);
 
         let mut snap_to_buildings = FindClosest::new();
         for b in app.map.all_buildings() {
@@ -80,7 +79,7 @@ impl Viewer {
 
         let start = app.map.get_b(b);
         self.isochrone = Isochrone::new(ctx, app, vec![start.id], self.isochrone.options.clone());
-        let star = draw_star(ctx, start);
+        let star = render::draw_star(ctx, start);
         self.highlight_start = ctx.upload(star);
         self.panel = build_panel(ctx, app, start, &self.isochrone);
         // Any previous hover is from the perspective of the old `highlight_start`.
@@ -217,12 +216,14 @@ impl State<App> for Viewer {
                 }
                 x => {
                     if let Some(category) = x.strip_prefix("businesses: ") {
-                        return Transition::Push(ExploreAmenities::new_state(
-                            ctx,
-                            app,
-                            &self.isochrone,
-                            AmenityType::from_str(category).unwrap(),
-                        ));
+                        return Transition::Push(
+                            crate::amenities_details::ExploreAmenitiesDetails::new_state(
+                                ctx,
+                                app,
+                                &self.isochrone,
+                                AmenityType::from_str(category).unwrap(),
+                            ),
+                        );
                     } else {
                         unreachable!()
                     }
@@ -233,7 +234,7 @@ impl State<App> for Viewer {
                     movement: options_from_controls(&self.panel),
                     thresholds: Options::default_thresholds(),
                 };
-                self.draw_unwalkable_roads = draw_unwalkable_roads(ctx, app, &options);
+                self.draw_unwalkable_roads = render::draw_unwalkable_roads(ctx, app, &options);
                 self.isochrone = Isochrone::new(ctx, app, vec![self.isochrone.start[0]], options);
                 self.panel = build_panel(
                     ctx,
@@ -313,12 +314,6 @@ fn options_from_controls(panel: &Panel) -> MovementOptions {
     } else {
         MovementOptions::Biking
     }
-}
-
-pub fn draw_star(ctx: &mut EventCtx, b: &Building) -> GeomBatch {
-    GeomBatch::load_svg(ctx, "system/assets/tools/star.svg")
-        .centered_on(b.polygon.center())
-        .color(RewriteColor::ChangeAll(Color::BLACK))
 }
 
 fn build_panel(ctx: &mut EventCtx, app: &App, start: &Building, isochrone: &Isochrone) -> Panel {
@@ -449,169 +444,4 @@ impl HoverOnBuilding {
             drawn_route: ctx.upload(batch),
         }
     }
-}
-
-struct ExploreAmenities {
-    table: Table<App, Entry, ()>,
-    panel: Panel,
-    draw: Drawable,
-}
-
-struct Entry {
-    bldg: BuildingID,
-    amenity_idx: usize,
-    name: String,
-    amenity_type: String,
-    address: String,
-    duration_away: Duration,
-}
-
-impl ExploreAmenities {
-    fn new_state(
-        ctx: &mut EventCtx,
-        app: &App,
-        isochrone: &Isochrone,
-        category: AmenityType,
-    ) -> Box<dyn State<App>> {
-        let mut batch = draw_isochrone(
-            &app.map,
-            &isochrone.time_to_reach_building,
-            &isochrone.thresholds,
-            &isochrone.colors,
-        );
-        batch.append(draw_star(ctx, app.map.get_b(isochrone.start[0])));
-
-        let mut entries = Vec::new();
-        for b in isochrone.amenities_reachable.get(category) {
-            let bldg = app.map.get_b(*b);
-            for (amenity_idx, amenity) in bldg.amenities.iter().enumerate() {
-                if AmenityType::categorize(&amenity.amenity_type) == Some(category) {
-                    entries.push(Entry {
-                        bldg: bldg.id,
-                        amenity_idx,
-                        name: amenity.names.get(app.opts.language.as_ref()).to_string(),
-                        amenity_type: amenity.amenity_type.clone(),
-                        address: bldg.address.clone(),
-                        duration_away: isochrone.time_to_reach_building[&bldg.id],
-                    });
-                    // Highlight the matching buildings
-                    batch.push(Color::RED, bldg.polygon.clone());
-                }
-            }
-        }
-
-        let mut table: Table<App, Entry, ()> = Table::new(
-            "time_to_reach_table",
-            entries,
-            // The label has extra junk to avoid crashing when one building has two stores,
-            // possibly with the same name in the current language
-            Box::new(|x| format!("{}: {} ({})", x.bldg.0, x.name, x.amenity_idx)),
-            "Time to reach",
-            Filter::empty(),
-        );
-        table.column(
-            "Type",
-            Box::new(|ctx, _, x| Text::from(&x.amenity_type).render(ctx)),
-            Col::Sortable(Box::new(|rows| {
-                rows.sort_by_key(|x| x.amenity_type.clone())
-            })),
-        );
-        table.static_col("Name", Box::new(|x| x.name.clone()));
-        table.static_col("Address", Box::new(|x| x.address.clone()));
-        table.column(
-            "Time to reach",
-            Box::new(|ctx, app, x| {
-                Text::from(x.duration_away.to_string(&app.opts.units)).render(ctx)
-            }),
-            Col::Sortable(Box::new(|rows| rows.sort_by_key(|x| x.duration_away))),
-        );
-
-        let panel = Panel::new_builder(Widget::col(vec![
-            Widget::row(vec![
-                Line(format!("{} within 15 minutes", category))
-                    .small_heading()
-                    .into_widget(ctx),
-                ctx.style().btn_close_widget(ctx),
-            ]),
-            table.render(ctx, app),
-        ]))
-        .aligned(HorizontalAlignment::Center, VerticalAlignment::TopInset)
-        .build(ctx);
-
-        Box::new(ExploreAmenities {
-            table,
-            panel,
-            draw: ctx.upload(batch),
-        })
-    }
-}
-
-impl State<App> for ExploreAmenities {
-    fn event(&mut self, ctx: &mut EventCtx, app: &mut App) -> Transition<App> {
-        ctx.canvas_movement();
-
-        match self.panel.event(ctx) {
-            Outcome::Clicked(x) => {
-                if self.table.clicked(&x) {
-                    self.table.replace_render(ctx, app, &mut self.panel)
-                } else if x == "close" {
-                    return Transition::Pop;
-                } else if let Some(idx) = x.split(':').next().and_then(|x| x.parse::<usize>().ok())
-                {
-                    let b = app.map.get_b(BuildingID(idx));
-                    open_browser(b.orig_id.to_string());
-                } else {
-                    unreachable!()
-                }
-            }
-            Outcome::Changed(_) => {
-                self.table.panel_changed(&self.panel);
-                self.table.replace_render(ctx, app, &mut self.panel)
-            }
-            _ => {}
-        }
-
-        Transition::Keep
-    }
-
-    fn draw(&self, g: &mut GfxCtx, app: &App) {
-        g.redraw(&self.draw);
-        self.panel.draw(g);
-        if let Some(x) = self
-            .panel
-            .currently_hovering()
-            .and_then(|x| x.split(':').next())
-            .and_then(|x| x.parse::<usize>().ok())
-        {
-            g.draw_polygon(Color::CYAN, app.map.get_b(BuildingID(x)).polygon.clone());
-        }
-    }
-}
-
-pub fn draw_unwalkable_roads(ctx: &mut EventCtx, app: &App, opts: &Options) -> Drawable {
-    let allow_shoulders = match opts.movement {
-        MovementOptions::Walking(ref opts) => opts.allow_shoulders,
-        MovementOptions::Biking => {
-            return Drawable::empty(ctx);
-        }
-    };
-
-    let mut batch = GeomBatch::new();
-    'ROADS: for road in app.map.all_roads() {
-        if road.is_light_rail() {
-            continue;
-        }
-        for l in &road.lanes {
-            if l.lane_type == LaneType::Sidewalk
-                || l.lane_type == LaneType::Footway
-                || l.lane_type == LaneType::SharedUse
-                || (l.lane_type == LaneType::Shoulder && allow_shoulders)
-            {
-                continue 'ROADS;
-            }
-        }
-        // TODO Skip highways
-        batch.push(Color::BLUE.alpha(0.5), road.get_thick_polygon());
-    }
-    ctx.upload(batch)
 }
